src/main/java/javacourses/Demo/Course.java
package javacourses.Demo;

public class Course {
	
	private String id;
	private String name;
	private String description;
	
	public Course(String id, String name, String description) {
		this.id = id;
		this.name = name;
		this.description = description;
	}
	
	/**
	 * @return the id
	 */
	public String getId() {
		return id;
	}


	/**
	 * @return the name
	 */
	public String getName() {
		return name;
	}


	/**
	 * @return the description
	 */
	public String getDescription() {
		return description;
	}

}
